This commit introduces significant changes to the `orderBy` method. Previously, the `orderBy` method required a direction parameter that was strictly bound to `ASC` or `DESC`.
With this update, the method can now accept a `null` direction, effectively ignoring the direction parameter. This change provides more flexibility when constructing queries, as it allows for ordering without specifying a direction, for example `RAND()`.

Additionally, this commit adds support for PostgreSQL-specific directions, namely `NULLS FIRST` and `NULLS LAST`. These options provide more control over the ordering of `null` values in the result set, which can be particularly useful when working with databases that contain nullable columns.